libc: make the <stdint.h> constant macros usable in #if
C11 7.20.4 requires INTN_C/UINTN_C/INTMAX_C/UINTMAX_C to expand to integer
constant expressions suitable for #if.  Substrate's wrapped each value in a
cast -- ((uint64_t) value ## ULL) -- and a cast is not a preprocessor
operator, so any #if that used one failed.  perl 5.44.0 does exactly that
through handy.h's nBIT_MASK, and dies building DynaLoader:

    handy.h:368: error: missing binary operator before token "1ULL"

Expand to the literal with its suffix only.  The suffix alone gives the
64-bit and unsigned 32-bit types; the 8- and 16-bit forms are plain int,
which is what the standard specifies after promotion.

Verified with the substrate cross gcc and host gcc -m32: #if checks on
UINT64_C(1) << 63, UINTMAX_C(0xFFFFFFFFFFFFFFFF), INT64_C(-1),
UINT32_C(0xFFFFFFFF), INT32_C(-5) and UINT8_C(255), plus sizeof checks
that the 64-bit forms are 8 bytes and UINT32_C is unsigned.  make -C tests
passes.

+++ include/stdint.h
@@ -123,20 +123,26 @@ typedef uint64_t uint_fast64_t;
 #define SIG_ATOMIC_MAX INT32_MAX
 
 /* Constants for minimum-width integer constant expressions */
-#define INT8_C(value) ((int8_t) value)
-#define INT16_C(value) ((int16_t) value)
-#define INT32_C(value) ((int32_t) value)
+/*
+ * C11 7.20.4: each expansion must be usable in #if, so no casts -- a cast is
+ * not a preprocessor operator, and `#if UINT64_C(1) << 63` would otherwise
+ * fail with "missing binary operator".  The suffix alone gives the type:
+ * the 8- and 16-bit forms promote to int exactly as the standard specifies.
+ */
+#define INT8_C(value)   value
+#define INT16_C(value)  value
+#define INT32_C(value)  value
 
-#define UINT8_C(value) ((uint8_t) value)
-#define UINT16_C(value) ((uint16_t) value)
-#define UINT32_C(value) ((uint32_t) value ## U)
+#define UINT8_C(value)  value
+#define UINT16_C(value) value
+#define UINT32_C(value) value ## U
 
 #if defined(__x86_64__) || defined(_M_X64)
-#define INT64_C(value) ((int64_t) value ## L)
-#define UINT64_C(value) ((uint64_t) value ## UL)
+#define INT64_C(value)  value ## L
+#define UINT64_C(value) value ## UL
 #else
-#define INT64_C(value) ((int64_t) value ## LL)
-#define UINT64_C(value) ((uint64_t) value ## ULL)
+#define INT64_C(value)  value ## LL
+#define UINT64_C(value) value ## ULL
 #endif
 
 #define INTMAX_C(value)  INT64_C(value)
